+++ src/candlewick/utils/LoadMaterial.h
@@ -5,6 +5,7 @@
 #include "Utils.h"
 
 #include <assimp/material.h>
+#include <string>
 
 namespace candlewick {
 
@@ -13,5 +14,9 @@ namespace candlewick {
 /// If the aiMaterial contains, in fact, a Phong material (\ref
 /// PhongMaterial), then a PBR material will be approximated.
 PbrMaterial loadFromAssimpMaterial(aiMaterial *material);
+
+/// \brief Extract the base color (or diffuse) texture path from an assimp
+/// material, if any.
+std::string getBaseColorTexturePath(aiMaterial *material);
 
 } // namespace candlewick